uses simplified resource interface for lifecycle resource
