+++ service/awsconfig/v10/drainer_resource_set.go
@@ -51,12 +51,7 @@ func NewDrainerResourceSet(config DrainerResourceSetConfig) (*framework.Resource
 			Service:   cloudFormationService,
 		}
 
-		ops, err := lifecycle.NewResource(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-
-		lifecycleResource, err = toCRUDResource(config.Logger, ops)
+		lifecycleResource, err = lifecycle.NewResource(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
