--- service/awsconfig/v10/resource/lifecycle/desired.go
@@ -1,9 +1,0 @@
-package lifecycle
-
-import (
-	"context"
-)
-
-func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interface{}, error) {
-	return nil, nil
-}
