--- service/awsconfig/v10/resource/lifecycle/update.go
@@ -1,19 +1,0 @@
-package lifecycle
-
-import (
-	"context"
-
-	"github.com/giantswarm/operatorkit/framework"
-)
-
-func (r *Resource) ApplyUpdateChange(ctx context.Context, obj, updateChange interface{}) error {
-	return nil
-}
-
-func (r *Resource) NewUpdatePatch(ctx context.Context, obj, currentState, desiredState interface{}) (*framework.Patch, error) {
-	return nil, nil
-}
-
-func (r *Resource) newUpdateChange(ctx context.Context, obj, currentState, desiredState interface{}) (interface{}, error) {
-	return nil, nil
-}
